I've refactored the dataset loading and manipulation utilities.
I moved dataset-related functions and utilities from the lmv3 module to a new, dedicated `src/dataset` directory.

This change makes these utilities more accessible for use by other modules in the future.

Specifically, I made the following changes:
- I created the `src/dataset` directory.
- I moved `filters.py`, `maps.py`, and `stats.py` from `src/lmv3` to `src/dataset`.
- I created `src/dataset/utils.py` and moved relevant functions (`load_labels`, `prepare_dataset`, `_to_fractional`) from `src/lmv3/utils/utils.py` into it.
- I updated all imports across the `src` directory to reflect the new locations of these files and functions.
- I ensured that `src/lmv3/utils/utils.py` and `src/lmv3/utils/wandb_utils.py` now use the common `_to_fractional` function from `src/dataset/utils.py`.
